Dont skip killer moves in qsearch bench 3954063
