+++ src/main/java/com/kelseyde/calvin/board/Board.java
@@ -395,6 +395,10 @@ public class Board {
         return pieces[square];
     }
 
+    public Piece captured(Move move) {
+        return move.isEnPassant() ? Piece.PAWN : pieceAt(move.to());
+    }
+
     public boolean isCapture(Move move) {
         return move.isEnPassant() || pieceAt(move.to()) != null;
     }
