--- src/main/java/com/kelseyde/calvin/search/picker/MoveType.java
@@ -1,9 +1,0 @@
-package com.kelseyde.calvin.search.picker;
-
-public enum MoveType {
-    TT_MOVE,
-    GOOD_NOISY,
-    KILLER,
-    QUIET,
-    BAD_NOISY
-}
