+++ src/main/java/com/kelseyde/calvin/search/picker/QuiescentMovePicker.java
@@ -7,18 +7,22 @@ import com.kelseyde.calvin.movegen.MoveGenerator;
 import com.kelseyde.calvin.movegen.MoveGenerator.MoveFilter;
 import com.kelseyde.calvin.search.SearchHistory;
 import com.kelseyde.calvin.search.SearchStack;
+import com.kelseyde.calvin.search.picker.AbstractMovePicker.Stage;
 
-public class QuiescentMovePicker extends MovePicker {
+import java.util.List;
+
+public class QuiescentMovePicker extends AbstractMovePicker {
 
     private MoveFilter filter;
 
+    private ScoredMove[] goodNoisies;
+
     public QuiescentMovePicker(
             EngineConfig config, MoveGenerator movegen, SearchStack ss, SearchHistory history, Board board, int ply, Move ttMove, boolean inCheck) {
-        super(config, movegen, ss, history, board, ply, ttMove, inCheck);
-        this.scorer.setSeeNoisyDivisor(config.seeQsNoisyDivisor());
-        this.scorer.setSeeNoisyOffset(config.seeQsNoisyOffset());
-        this.stage = ttMove != null ? Stage.TT_MOVE : Stage.QSEARCH_GEN_NOISY;
-        this.skipQuiets = true;
+        super(movegen,
+                new MoveScorer(config, history, ss, config.seeQsNoisyDivisor(), config.seeQsNoisyOffset()),
+                history, board, ply, ttMove, inCheck);
+        this.stage = Stage.TT_MOVE;
     }
 
     @Override
@@ -30,12 +34,36 @@ public class QuiescentMovePicker extends MovePicker {
                 case TT_MOVE -> pickTTMove(Stage.QSEARCH_GEN_NOISY);
                 case QSEARCH_GEN_NOISY -> generate(filter, Stage.QSEARCH_NOISY);
                 case QSEARCH_NOISY -> pickMove(Stage.END);
-                case GEN_NOISY, GOOD_NOISY, KILLER, GEN_QUIET, QUIET, BAD_NOISY, END -> null;
+                default -> null;
             };
             if (stage == Stage.END) break;
         }
         return nextMove;
 
+    }
+
+    @Override
+    protected void handleStagedMoves(List<Move> moves) {
+        // In quiescent movegen all moves are treated as 'good noisies'
+        goodNoisies = new ScoredMove[moves.size()];
+        int goodIndex = 0;
+        for (Move move : moves) {
+            ScoredMove scoredMove = scorer.score(board, move, ply, stage);
+            // In q-search, only consider good noisies
+            // unless we are in check, in which case consider all moves.
+            if (scoredMove.isGoodNoisy() || inCheck)
+                goodNoisies[goodIndex++] = scoredMove;
+        }
+    }
+
+    @Override
+    protected boolean isSpecial(Move move) {
+        return move.equals(ttMove)
+    }
+
+    @Override
+    protected ScoredMove[] loadStagedMoves(Stage stage) {
+        return goodNoisies;
     }
 
     public void setFilter(MoveFilter filter) {
