+++ src/main/java/com/kelseyde/calvin/search/picker/ScoredMove.java
@@ -8,26 +8,26 @@ public record ScoredMove(Move move,
                          Piece captured,
                          int score,
                          int historyScore,
-                         MoveType moveType) {
+                         AbstractMovePicker.MoveType moveType) {
 
     public boolean isNoisy() {
-        return moveType == MoveType.GOOD_NOISY || moveType == MoveType.BAD_NOISY;
+        return moveType == AbstractMovePicker.MoveType.GOOD_NOISY || moveType == AbstractMovePicker.MoveType.BAD_NOISY;
     }
 
     public boolean isGoodNoisy() {
-        return moveType == MoveType.GOOD_NOISY;
+        return moveType == AbstractMovePicker.MoveType.GOOD_NOISY;
     }
 
     public boolean isKiller() {
-        return moveType == MoveType.KILLER;
+        return moveType == AbstractMovePicker.MoveType.KILLER;
     }
 
     public boolean isBadNoisy() {
-        return moveType == MoveType.BAD_NOISY;
+        return moveType == AbstractMovePicker.MoveType.BAD_NOISY;
     }
 
     public boolean isQuiet() {
-        return moveType == MoveType.QUIET;
+        return moveType == AbstractMovePicker.MoveType.QUIET;
     }
 
 }
